refactor: Updated file downloading import
